Moves packages around a bit
