--- evaluation/object/constants.go
@@ -1,5 +1,0 @@
-package object
-
-var (
-
-)
