+++ object/compound.go
